verdict-pdf-service/requirements.txt
# requirements.txt
fastapi
uvicorn
fpdf2
python-dotenv
supabase-py
